+++ CHANGELOG.md
@@ -151,8 +151,16 @@ prompt is displayed.
     - Added `Cmd2ArgumentParser.output_to()` context manager to temporarily set the output stream
       during `argparse` operations. This is helpful for directing output for functions like
       `parse_args()`, which default to `sys.stdout` and lack a `file` argument.
-    - Added ability to customize `prompt-toolkit` completion menu colors by overriding
-      `Cmd2Style.COMPLETION_MENU_ITEM` and `Cmd2Style.COMPLETION_MENU_META` in the `cmd2` theme.
+    - Added ability to customize `prompt-toolkit` completion menu colors by overriding the following
+      fields in the `cmd2` theme:
+        - `Cmd2Style.COMPLETION_MENU` - Base style for the entire completion menu container (sets
+          the background)
+        - `Cmd2Style.COMPLETION_MENU_COMPLETION` -Style for an individual, non-selected completion
+          item
+        - `Cmd2Style.COMPLETION_MENU_CURRENT` - Style for the currently selected completion item
+        - `Cmd2Style.COMPLETION_MENU_META` - Style for "meta" information shown alongside a
+          completion
+        - `Cmd2Style.COMPLETION_MENU_META_CURRENT`- Style for meta info of current item
 
 ## 3.5.1 (April 24, 2026)